# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/{vidyaId}:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: The user's notification feed
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notifications, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
  /notifications/{vidyaId}/read:
    post:
      operationId: markNotificationsRead
      tags: [notifications]
      summary: Mark notifications as read
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadInput"
      responses:
        "200":
          description: Updated feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
...
  /tutor/{vidyaId}/assignments:
    get:
      operationId: listTutorAssignments
      tags: [tutor]
      summary: Assignments the tutor has created
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tutor assignments with completion counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorAssignmentsResponse"
    post:
      operationId: createAssignment
      tags: [tutor]
      summary: Assign work to a batch
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentInput"
      responses:
        "200":
          description: Updated tutor assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorAssignmentsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tutor/{vidyaId}/announcements:
    get:
      operationId: listAnnouncements
      tags: [tutor]
      summary: Announcements the tutor has sent
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Announcements, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementsResponse"
    post:
      operationId: createAnnouncement
      tags: [tutor]
      summary: Broadcast a message to a batch
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementInput"
      responses:
        "200":
          description: Updated announcements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tutor/{vidyaId}/heatmap:
    get:
      operationId: getClassHeatmap
      tags: [tutor]
      summary: Per-topic mastery across a batch
      parameters:
        - in: path
          name: vidyaId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Class heatmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeatmapResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    NotificationItem:
      type: object
      required: [id, type, title, body, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: string
        linkTab:
          type: ["string", "null"]
        read:
          type: boolean
        createdAt:
          type: string
    NotificationsResponse:
      type: object
      required: [unreadCount, notifications]
      properties:
        unreadCount:
          type: integer
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationItem"
    MarkReadInput:
      type: object
      properties:
        ids:
          type: ["array", "null"]
          items:
            type: integer
        all:
          type: ["boolean", "null"]
...
    CreateAssignmentInput:
      type: object
      required: [batch, title, kind]
      properties:
        batch:
          type: string
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        kind:
          type: string
          enum: [practice, test]
        topic:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
    TutorAssignment:
      type: object
      required:
        [id, batch, title, kind, createdAt, completedCount, totalCount]
      properties:
        id:
          type: integer
        batch:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        kind:
          type: string
        topic:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        completedCount:
          type: integer
        totalCount:
          type: integer
    TutorAssignmentsResponse:
      type: object
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/TutorAssignment"
    Announcement:
      type: object
      required: [id, batch, message, createdAt]
      properties:
        id:
          type: integer
        batch:
          type: string
        message:
          type: string
        createdAt:
          type: string
    AnnouncementsResponse:
      type: object
      required: [announcements]
      properties:
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
    CreateAnnouncementInput:
      type: object
      required: [batch, message]
      properties:
        batch:
          type: string
        message:
          type: string
          minLength: 1
    HeatmapRow:
      type: object
      required: [studentVidyaId, name, batch, topics]
      properties:
        studentVidyaId:
          type: string
        name:
          type: string
        batch:
          type: ["string", "null"]
        topics:
          type: object
          additionalProperties:
            type: integer
    HeatmapResponse:
      type: object
      required: [topics, rows]
      properties:
        topics:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/HeatmapRow"
...