# lib/api-zod/src/generated/types/heatmapResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeatmapRow } from './heatmapRow';

export interface HeatmapResponse {
  topics: string[];
  rows: HeatmapRow[];
}