# lib/api-zod/src/generated/types/tutorAssignment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TutorAssignment {
  id: number;
  batch: string;
  title: string;
  /** @nullable */
  description?: string | null;
  kind: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  completedCount: number;
  totalCount: number;
}