# lib/api-zod/src/generated/types/tutorAssignmentsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TutorAssignment } from './tutorAssignment';

export interface TutorAssignmentsResponse {
  assignments: TutorAssignment[];
}